+++ src/templates/export.js
@@ -149,6 +149,8 @@ module.exports = (router) => {
             'properties',
             'controller',
             'submissionRevisions',
+            'revisions',
+            'esign',
             ...includeFormFields,
           );
           if (form.revisions) {
@@ -174,6 +176,16 @@ module.exports = (router) => {
             }
             if (component && component.project) {
               component.project = 'project';
+            }
+            // During export if select component
+            // data type resource de-ref defaultValue
+            if (
+              component
+              && component.type === 'select'
+              && component.dataSrc === 'resource'
+              && component.defaultValue
+            ) {
+              component.defaultValue = undefined;
             }
 
             if (component.hasOwnProperty('form') && component.revision) {
@@ -292,6 +304,7 @@ module.exports = (router) => {
                     'properties',
                     'controller',
                     'submissionRevisions',
+                    'esign',
                     '_vid',
                     'revisionId',
                     ...includeFormFields